Fix unnecessary DNAT to improve pihole logs
We were DNATing every packet, but most are already being sent to the pihole in the first place,
so this is unnecessary. For traffic from the other VLAN, in particular, this was causing everything
to be masqueraded in the POSTROUTING chain.

There are two changes here: do not DNAT traffic which is going to the pihole, and do not masquerade
traffic which had not been DNATed. The result is that "legit" DNS traffic shows up correctly in the logs,
whereas traffic which was intended for another DNS server shows up as the gateway.